% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtMonthlyMacroData}
\alias{dtMonthlyMacroData}
\title{Selected monthly macro time series}
\format{
data.table, 36 columns, 289 rows
}
\source{
\url{https://stat.gov.pl/}, \url{https://www.nbp.pl/}
}
\usage{
dtMonthlyMacroData
}
\description{
This data set has been compiled using GUS data and NBP data.
}
\keyword{datasets}
