% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtQuarterlyMacroData}
\alias{dtQuarterlyMacroData}
\title{Selected quarterly macro time series}
\format{
data.table, 17 columns, 83 rows
}
\source{
\url{https://stat.gov.pl/}, \url{https://www.nbp.pl/}
}
\usage{
dtQuarterlyMacroData
}
\description{
This data set has been compiled using GUS data and NBP data.
}
\keyword{datasets}
